README.md
This is the report of my analysis of Net-Super Store, a fast-moving consumer goods organisation (FMCG) with a focus on the distribution & sales of food items, furniture, clothing and electronics. This analysis is to analyse the sales data from the store in other to uncover the sales trend and how to help this business improve, some of the data needed for this analysis include; product id, sales rep, region, product category, customer type, payment method, region sales rep, sales channel, quantity sold, sales date, sales amount, unit cost, unit cost and discount.